Safari-611.1.10/LayoutTests/accessibility/mac/caret-browsing-attribute.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="resources/accessibility-helper.js"></script>
</head>
<body id="body">
<script>
description("This tests setting caret browsing preference via ax api.");
if (window.testRunner) {

  testRunner.dumpAsText();

  if (window.accessibilityController) {

    var webArea = accessibilityController.rootElement.childAtIndex(0);

    setCaretBrowsingEnabled(webArea, true);
    shouldBe("caretBrowsingEnabled(webArea)", "true");

    setCaretBrowsingEnabled(webArea, false);
    shouldBe("caretBrowsingEnabled(webArea)", "false");
  }
}
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
