Safari-611.1.10/LayoutTests/accessibility/mac/caret-browsing-attribute-expected.txt

This tests setting caret browsing preference via ax api.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretBrowsingEnabled(webArea) is true
PASS caretBrowsingEnabled(webArea) is false
PASS successfullyParsed is true

TEST COMPLETE
